Update Authentication_with_OAuth.ipynb
The equal sign after --scopes and the single quotes before the urls were being interpreted as part of the first scope URL. The last single quote was being interpreted as part of the third scope URL.

This caused a 400 error with the message "Some scopes invalid" and prevented sign in. Running the command without the = or ' was successful.