Use non-experimental model name for 2.0
Updated the flash-exp model references to drop the `-exp` suffix. Also
switched each of these to use a model selector where it wasn't already
configured.

I haven't added any extra models to notebooks with the selector as I
figure they should be tested, and I didn't do that, but we can do it in
future updates if we wish.